Switch call for kindIs invalid to empty since value will always be provided
In rancher-project-monitoring, a dummy empty string (not null) value will always be provided for both `global.cattle.clusterId` and `global.cattle.url` by the Helm Project Operator:

https://github.com/rancher/helm-project-operator/blob/557114ccbc4137b9272a54705416b03043f6a11a/pkg/controllers/project/values.go#L14

https://github.com/rancher/helm-project-operator/blob/557114ccbc4137b9272a54705416b03043f6a11a/pkg/controllers/project/values.go#L29

Therefore, the `{{ kindIs "invalid" FIELD }}` call is failing here since the kind is not invalid; both are of type string. Instead, `empty` should be used for the check here.