Ensure release namespace is always included for dashboards and metrics
By default, Helm Project Operator only targets namespaces that are already label-selected into a project unless it has been created by the operator itself (when it meets the conditions to create a separate release namespace), as seen in:

https://github.com/rancher/helm-project-operator/blob/72233daffbda6d59c2745f4697e633bf90e591f1/pkg/controllers/project/controller.go#L248-L254

However, when a user is not using an auto-generated separate release namespace, the Project Prometheus Stack will fail to monitor anything since the release namespace resources (i.e. default alerts and dashboards) are not picked up unless the release namespace has the same project annotation.

Therefore, this fix ensures that we always monitor the release namespace no matter what by adding it into the projectReleaseNamespace list when it does not exist.