Add ability to determine if PromFed should manage HelmController CRDs
For now we indirectly rely on checking nodes for k8s runtime type, then based on if we find k3s/rke2 we skip them.
In the future, we'd like to directly be able to query the CRD on the cluster and see a clear annotation/label citing the CRDs owner.